requirements.txt
accelerate==0.30.1
alembic==1.13.1
altair==5.4.1
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
apturl==0.5.2
asgiref==3.5.0
attr==0.3.1
attrs==21.2.0
azure-core==1.30.1
azure-storage-blob==12.19.1
bcrypt==4.1.2
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==5.0.1
blinker==1.4
boto==2.49.0
boto3==1.34.97
botocore==1.34.97
boxing==0.1.4
Brlapi==0.8.3
Brotli==1.0.9
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
colour==0.1.5
command-not-found==0.3
cryptography==42.0.5
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
diffusers==0.27.2
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==3.2.25
django-annoying==0.10.6
django-cors-headers==3.6.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.13.1
dnspython==2.6.1
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
ecdsa==0.19.0
email_validator==2.1.1
exceptiongroup==1.2.0
expiringdict==1.2.2
fastapi==0.110.1
filelock==3.13.4
fonttools==4.29.1
fs==2.4.12
fsspec==2024.3.1
gast==0.5.2
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-appengine-logging==1.4.3
google-cloud-audit-log==0.2.5
google-cloud-core==2.4.1
google-cloud-logging==3.10.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.63.0
grpcio-status==1.62.2
gyp==0.1
h11==0.14.0
html5lib==1.1
htmlmin==0.1.12
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.0
humansignal-drf-yasg==1.21.9
idna==3.6
ijson==3.2.3
imgviz==1.4.1
importlib-metadata==4.6.4
imutils==0.5.4
inflection==0.5.1
iniconfig==2.0.0
isodate==0.6.1
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.3.2
label-studio==1.12.0
label-studio-converter==0.0.58
label-studio-tools==0.0.4
labelImg==1.8.6
labelme==4.6.0
language-selector==0.1
launchdarkly-server-sdk==8.2.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.5.1
mdurl==0.1.2
mock==4.0.3
more-itertools==8.10.0
mpmath==0.0.0
narwhals==1.8.3
netifaces==0.11.0
networkx==3.3
nltk==3.6.7
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
openai==1.25.1
opencv-python==4.9.0.80
ordered-set==4.0.2
orjson==3.10.0
packaging==21.3
pandas==2.2.2
passlib==1.7.4
pbr==5.8.0
pexpect==4.8.0
pillow==10.3.0
platformdirs==4.2.1
pluggy==1.4.0
ply==3.11
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.6
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==1.10.15
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
pydeck==0.9.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyQtWebEngine==5.15.5
pyRFC3339==1.1
pyrsistent==0.20.0
pytest==8.1.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==2.0.4
python-multipart==0.0.9
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==6.0.1
QtPy==2.0.0
redis==3.5.3
regex==2024.4.28
reportlab==3.6.8
requests==2.31.0
rich==13.8.1
rq==1.10.1
rsa==4.9
rules==2.2
s3transfer==0.10.1
safetensors==0.4.3
scipy==1.8.0
screen-resolution-extra==0.0.0
seaborn==0.13.2
SecretStorage==3.3.1
semver==3.0.2
sentry-sdk==2.0.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.3.1
SQLAlchemy==2.0.29
sqlparse==0.5.0
starlette==0.37.2
streamlit==1.38.0
sympy==1.9
systemd-python==234
tenacity==8.5.0
termcolor==1.1.0
thop==0.1.1.post2209072238
toml==0.10.2
tomli==2.0.1
torch==2.2.2
torchvision==0.17.2
tornado==6.4.1
tqdm==4.66.2
triton==2.2.0
typing_extensions==4.11.0
tzdata==2024.1
ua-parser==0.18.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
ujson==5.9.0
unattended-upgrades==0.1
unicodedata2==14.0.0
uritemplate==4.1.1
urllib3==1.26.18
user-agents==2.2.0
uvicorn==0.29.0
uvloop==0.19.0
virtualenv==20.26.1
wadllib==1.3.6
watchdog==4.0.2
watchfiles==0.21.0
webencodings==0.5.1
websockets==12.0
wsproto==1.0.0
xdg==5
xkit==0.0.0
xmljson==0.2.0
zipp==1.0.0
